sample_work_project/requirements.txt
amazoncaptcha==0.5.11
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
comm==0.2.1
cryptography==41.0.3
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
EZGmail==2022.10.10
fastjsonschema==2.19.1
fqdn==1.5.1
google-api-core==2.13.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.4
ipykernel==6.29.2
ipython==8.22.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.3
oauth2client==4.1.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
paramiko==3.3.1
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
schedule==1.2.1
selenium==4.15.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
trio==0.23.1
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.19
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
wsproto==1.2.0
zi-api-auth-client==2.0.1
